# src/main/java/com/BiteBooking/backend/model/RatingImage.java
package com.BiteBooking.backend.model;

import com.fasterxml.jackson.annotation.JsonBackReference;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Entity
@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@Table(name = "rating_images")
public class RatingImage {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String imagePath;

    @Column(nullable = false)
    private Integer imageOrder;

    @ManyToOne
    @JoinColumn(name = "rating_id", nullable = false)
    @JsonBackReference
    private Rating rating;
}